move validators to own file, validate and colour formula and adduct
